Added supabase to contact form
