+++ package.json
@@ -15,6 +15,7 @@
     "@radix-ui/react-slot": "^1.2.2",
     "@react-three/drei": "^10.2.0",
     "@react-three/fiber": "^9.0.0-alpha.8",
+    "@supabase/auth-helpers-nextjs": "^0.10.0",
     "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "latest",
     "@tabler/icons-react": "^3.34.0",
